crypto-monitoring/requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
kafka-python==2.2.15
py4j==0.10.9.7
pyspark==3.4.1
requests==2.32.5
urllib3==2.5.0
influxdb-client==1.48.0
python-dotenv==1.0.1
tenacity==9.1.2
pybreaker==1.4.1
avro-python3==1.10.2
websocket-client==1.9.0
feedparser==6.0.11
